+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Removed
 - ``Slick\Template\EngineInterface`` it was deprecated and is removed.
+- ``Slick\Exception\ParserException`` it wold throw source engine exceptions instead.
+- ``Slick\Common`` library dependency
 
 ## [v1.2.5] - 2015-05-09